+++ lencod/src/cabac.c
@@ -707,7 +707,7 @@ void writeRefFrame_CABAC(SyntaxElement *se, EncodingEnvironmentPtr eep_dp)
   int   a, b;
   int   act_ctx;
   int   act_sym;
-  int** refframe_array = enc_picture->ref_idx[se->value2];
+  short** refframe_array = enc_picture->ref_idx[se->value2];
 
   int bslice = (img->type==B_SLICE);
